+++ IceCream/Classes/CKRecordConvertible.swift
@@ -19,45 +19,6 @@ public protocol CKRecordConvertible {
     var isDeleted: Bool { get }
 }
 
-public protocol CKRecordRecoverable {
-    
-}
-
-extension CKRecordRecoverable where Self: Object {
-    func parseFromRecord(record: CKRecord) -> Self? {
-        let o = Self()
-        for prop in o.objectSchema.properties {
-            var recordValue: Any?
-            switch prop.type {
-            case .int:
-                recordValue = record.value(forKey: prop.name) as? Int
-            case .string:
-                recordValue = record.value(forKey: prop.name) as? String
-            case .bool:
-                recordValue = record.value(forKey: prop.name) as? Bool
-            case .date:
-                recordValue = record.value(forKey: prop.name) as? Date
-            case .float:
-                recordValue = record.value(forKey: prop.name) as? Float
-            case .double:
-                recordValue = record.value(forKey: prop.name) as? Double
-            case .data:
-                recordValue = record.value(forKey: prop.name) as? Data
-            case .object:
-                guard let asset = record.value(forKey: prop.name) as? CKAsset else {
-                    print("For now, the Object only support CKAsset related type.")
-                    break
-                }
-                recordValue = CreamAsset.parse(from: prop.name, record: record, asset: asset)
-            default:
-                print("Other types will be supported in the future.")
-            }
-            o.setValue(recordValue, forKey: prop.name)
-        }
-        return o
-    }
-}
-
 extension CKRecordConvertible where Self: Object {
     
     public static var recordType: String {
